+++ src/main/webapp/resources/js/pages/projects/ProjectSPA.tsx
@@ -1,15 +1,23 @@
-import React, {Suspense} from "react";
-import {DataBrowserRouter, Outlet, Route} from "react-router-dom";
-import {loader as exportsLoader,} from "../../components/ncbi/export-table";
-import {setBaseUrl} from "../../utilities/url-utilities";
-import ProjectNCBILayout from "./ncbi";
-import {loader as detailsLoader,} from "../../components/ncbi/details";
-import {LoadingOutlined} from "@ant-design/icons";
+import React, { Suspense } from "react";
+import { render } from "react-dom";
+import { DataBrowserRouter, Outlet, Route } from "react-router-dom";
+import { loader as exportsLoader } from "../../components/ncbi/export-table";
+import { setBaseUrl } from "../../utilities/url-utilities";
+import { loader as detailsLoader } from "../../components/ncbi/details";
+import { LoadingOutlined } from "@ant-design/icons";
+import { loader as ncbiLoader } from "./ncbi/create";
 
-
-const NCBIExportDetails = React.lazy(() => import("../../components/ncbi/details"));
-const NcbiExportTable = React.lazy(() => import("../../components/ncbi/export-table"));
-const DefaultErrorBoundary = React.lazy(() => import ("../../components/DefaultErrorBoundary"));
+const ProjectNCBILayout = React.lazy(() => import("./ncbi"));
+const NCBIExportDetails = React.lazy(
+  () => import("../../components/ncbi/details")
+);
+const NcbiExportTable = React.lazy(
+  () => import("../../components/ncbi/export-table")
+);
+const NcbiCreateExport = React.lazy(() => import("./ncbi/create"));
+const DefaultErrorBoundary = React.lazy(
+  () => import("../../components/DefaultErrorBoundary")
+);
 
 __webpack_public_path__ = setBaseUrl(`/dist/`);
 
@@ -21,7 +29,9 @@ function ProjectBase(): JSX.Element {
   return (
     <div>
       {/* TODO: NAV AND OTHER TOP LEVEL ITEMS HERE */}
-        <Suspense fallback={<LoadingOutlined />}><Outlet/></Suspense>
+      <Suspense fallback={<LoadingOutlined />}>
+        <Outlet />
+      </Suspense>
     </div>
   );
 }
@@ -39,7 +49,17 @@ export default function ProjectSPA(): JSX.Element {
         path={setBaseUrl(`/projects/:projectId`)}
         element={<ProjectBase />}
       >
-        <Route path="export" element={<ProjectNCBILayout />}>
+        <Route
+          path="ncbi"
+          element={<NcbiCreateExport />}
+          loader={ncbiLoader}
+          errorElement={<DefaultErrorBoundary />}
+        />
+        <Route
+          path="export"
+          element={<ProjectNCBILayout />}
+          errorElement={<DefaultErrorBoundary />}
+        >
           <Route
             index
             element={<NcbiExportTable />}
@@ -57,3 +77,5 @@ export default function ProjectSPA(): JSX.Element {
     </DataBrowserRouter>
   );
 }
+
+render(<ProjectSPA />, document.querySelector("#root"));