+++ src/test/java/ca/corefacility/bioinformatics/irida/web/controller/test/integration/util/ITestAuthUtils.java
@@ -6,13 +6,12 @@ import static io.restassured.path.json.JsonPath.from;
 import java.util.HashMap;
 import java.util.Map;
 
+import io.restassured.path.json.exception.JsonPathException;
 import io.restassured.response.Response;
 import io.restassured.specification.RequestSpecification;
 
 /**
  * Utilities for doing web requests as certain types of user roles.
- * 
- * 
  */
 public class ITestAuthUtils {
 
@@ -25,11 +24,11 @@ public class ITestAuthUtils {
 	private static final String ROLE_OTHER_USER = "other";
 	public static final String BAD_USERNAME = "bad_username";
 	public static final String BAD_PASSWORD = "bad_password";
-	
+
 	private static final String CLIENT_ID = "testClient";
 	private static final String CLIENT_SECRET = "testClientSecret";
 	private static final String OAUTH_ENDPOINT = "/api/oauth/token";
-	
+
 	static {
 		ROLE_TO_USER = new HashMap<>();
 		ROLE_TO_USER.put(ROLE_ADMIN, new AuthenticationHolder("admin", "password1"));
@@ -43,30 +42,35 @@ public class ITestAuthUtils {
 
 	public static RequestSpecification asRole(String role) {
 		AuthenticationHolder pair = ROLE_TO_USER.get(role);
-		String oAuthToken = getOAuthToken(pair.username,pair.password);
+		String oAuthToken = getOAuthToken(pair.username, pair.password);
 		pair.setToken(oAuthToken);
 		String authString = "Bearer " + oAuthToken;
-		return given().header("Authorization", authString);			
+		return given().header("Authorization", authString);
 	}
-	
+
 	public static String getTokenForRole(final String role) {
-		AuthenticationHolder pair = ROLE_TO_USER.get(role);		
+		AuthenticationHolder pair = ROLE_TO_USER.get(role);
 		if (pair.token == null) {
 			asRole(role);
 		}
 		return pair.token;
 	}
-	
-	private static String getOAuthToken(String username, String password){
-		Response response = given().param("grant_type", "password")
-			.param("client_id", CLIENT_ID)
-			.param("client_secret", CLIENT_SECRET)
-			.param("username", username)
-			.param("password", password)
-			.get(OAUTH_ENDPOINT);
-		
-		String token = from(response.getBody().asString()).getString("access_token");
-		return token;
+
+	private static String getOAuthToken(String username, String password) {
+		Response response = given().auth()
+				.preemptive()
+				.basic(CLIENT_ID, CLIENT_SECRET)
+				.contentType("application/x-www-form-urlencoded")
+				.formParam("grant_type", "password")
+				.formParam("username", username)
+				.formParam("password", password)
+				.post(OAUTH_ENDPOINT);
+		try {
+			String token = from(response.getBody().asString()).getString("access_token");
+			return token;
+		} catch (JsonPathException ex) {
+			return "";
+		}
 	}
 
 	/**
@@ -77,7 +81,7 @@ public class ITestAuthUtils {
 	public static RequestSpecification asUser() {
 		return asRole(ROLE_USER);
 	}
-	
+
 	/**
 	 * Execute an HTTP request as a *different* user.
 	 * 
@@ -104,7 +108,7 @@ public class ITestAuthUtils {
 	public static RequestSpecification asAdmin() {
 		return asRole(ROLE_ADMIN);
 	}
-	
+
 	/**
 	 * Execute an HTTP request as a sequencer.
 	 * 
@@ -123,7 +127,7 @@ public class ITestAuthUtils {
 			this.username = username;
 			this.password = password;
 		}
-		
+
 		public void setToken(final String token) {
 			this.token = token;
 		}